# ACL/src/main/resources/static/swagger.yml
swagger: '2.0'
info:
  description: security API documentation
  version: 0.0.1
  title: security API
  contact: {}
  license:
    name: unlicensed
host: localhost:8085
basePath: /
...
paths:
...
  /api/acl-object-identities:
    get:
      tags:
        - acl-object-identity-resource
      summary: getAllAclObjectIdentities
      operationId: getAllAclObjectIdentitiesUsingGET
      produces:
        - '*/*'
      parameters:
        - name: page
          in: query
          description: Page number of the requested page
          required: false
          type: integer
          format: int32
        - name: size
          in: query
          description: Size of a page
          required: false
          type: integer
          format: int32
        - name: sort
          in: query
          description: 'Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported.'
          required: false
          type: array
          items:
            type: string
          collectionFormat: multi
        - name: X-TENANT-ID
          in: header
          description: X-TENANT-ID
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/AclObjectIdentityDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - acl-object-identity-resource
      summary: createAclObjectIdentity
      operationId: createAclObjectIdentityUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: aclObjectIdentityDTO
          description: aclObjectIdentityDTO
          required: true
          schema:
            $ref: '#/definitions/AclObjectIdentityDTO'
        - name: X-TENANT-ID
          in: header
          description: X-TENANT-ID
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/AclObjectIdentityDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /api/acl-object-identities/{id}:
    get:
      tags:
        - acl-object-identity-resource
      summary: getAclObjectIdentity
      operationId: getAclObjectIdentityUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
        - name: X-TENANT-ID
          in: header
          description: X-TENANT-ID
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/AclObjectIdentityDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
        - acl-object-identity-resource
      summary: updateAclObjectIdentity
      operationId: updateAclObjectIdentityUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: aclObjectIdentityDTO
          description: aclObjectIdentityDTO
          required: true
          schema:
            $ref: '#/definitions/AclObjectIdentityDTO'
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
        - name: X-TENANT-ID
          in: header
          description: X-TENANT-ID
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/AclObjectIdentityDTO'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - acl-object-identity-resource
      summary: deleteAclObjectIdentity
      operationId: deleteAclObjectIdentityUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
        - name: X-TENANT-ID
          in: header
          description: X-TENANT-ID
          required: false
          type: string
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    patch:
      tags:
        - acl-object-identity-resource
      summary: partialUpdateAclObjectIdentity
      operationId: partialUpdateAclObjectIdentityUsingPATCH
      consumes:
        - application/merge-patch+json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: aclObjectIdentityDTO
          description: aclObjectIdentityDTO
          required: true
          schema:
            $ref: '#/definitions/AclObjectIdentityDTO'
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
        - name: X-TENANT-ID
          in: header
          description: X-TENANT-ID
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/AclObjectIdentityDTO'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
definitions:
  AclClassDTO:
    type: object
    required:
      - classIdType
      - className
    properties:
      classIdType:
        type: string
      className:
        type: string
      id:
        type: integer
        format: int64
    title: AclClassDTO
...
  AclObjectIdentityDTO:
    type: object
    required:
      - objectIdIdentity
      - parentObject
    properties:
      aclClass:
        $ref: '#/definitions/AclClassDTO'
      entriesInheriting:
        type: boolean
      id:
        type: integer
        format: int64
      objectIdIdentity:
        type: string
      ownerSid:
        type: integer
        format: int32
      parentObject:
        type: integer
        format: int32
    title: AclObjectIdentityDTO
...